Added MATLAB to Simulink time array
BinaryArray takes the inputs: phone number, duration, silence
Outputs to a .MAT file for Simulink to read

ReadTimeSeries - is a basic function to read the MAT file. Here you will
see 1 or 0 depending in when the frequency should be on or off.

SineWaves -
Inputs the .MAT created by BinaryArray.m
Outputs the signal in Simulink